pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-manager"
version = "1.0.0"
authors = [
    {name = "Tu Nombre", email = "tu@email.com"},
]
description = "GUI Manager for ComfyUI with system monitoring and optimizations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "Pillow>=10.0.0",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
gpu = ["nvidia-ml-py>=11.0.0", "GPUtil>=1.4.0"]
dev = ["black>=23.0.0", "flake8>=6.0.0", "pytest>=7.0.0", "pytest-qt>=4.0.0"]
themes = ["ttkthemes>=3.2.2"]

[project.urls]
Homepage = "https://github.com/tuusuario/comfyui-manager"
Bug-Tracker = "https://github.com/tuusuario/comfyui-manager/issues"

[project.scripts]
comfyui-manager = "comfyui_manager.main:main"

[tool.setuptools.packages.find]
where = ["src"]